Inhertance structure
Add Account class
