+++ app/src/main/java/com/example/niot/deliveryfood/User.java
@@ -5,22 +5,7 @@ import com.google.gson.annotations.SerializedName;
 
 import java.io.Serializable;
 
-public class User implements Serializable {
-    @SerializedName("id_nd")
-    @Expose
-    private int id;
-    @SerializedName("ten_nd")
-    @Expose
-    private String name;
-    @SerializedName("email_nd")
-    @Expose
-    private String email;
-    @SerializedName("sdt_nd")
-    @Expose
-    private String phone;
-    @SerializedName("password_nd")
-    @Expose
-    private String password;
+public class User extends Account implements Serializable {
     @SerializedName("dchi_nd")
     @Expose
     private String address;
@@ -30,26 +15,6 @@ public class User implements Serializable {
     @SerializedName("is_login_nd")
     @Expose
     private boolean isLogin;
-
-    public int getId(){
-        return id;
-    }
-
-    public String getName(){
-        return name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public String getPhone() {
-        return phone;
-    }
-
-    public String getPassword() {
-        return password;
-    }
 
     public String getAddress() {
         return address;
@@ -65,8 +30,6 @@ public class User implements Serializable {
 
     @Override
     public String toString() {
-        String ans = "ID = " + String.valueOf(id) + "\nName = " + name + "\nEmail = "
-                + email + "\nAddress = " + address + "\nPhone = " + phone;
-        return ans;
+        return super.toString() + "\nAddress: " + address + "\nisLogin: " + isLogin;
     }
 }
